High Sierra Fixes

+++ use.sh
@@ -65,26 +65,59 @@ snap_classic_install ${snap_classic_pkgs[@]}
 
 if [ "$(uname)" == "Darwin" ]; then
   # install developer tools
-  touch /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress;
-  PROD=$(softwareupdate -l |
-  grep "\*.*Command Line.*$(sw_vers -productVersion|awk -F. '{print $1"."$2}')" |
-  head -n 1 | awk -F"*" '{print $2}' |
-  sed -e 's/^ *//' |
-  tr -d '\n')
-  softwareupdate -i "$PROD" --verbose
-  rm /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+  version=$(sw_vers -productVersion|awk -F. '{print $1"."$2}')
+  if [ version == "10.9" ]; then
+    chk_cmd="pkgutil --pkg-info=com.apple.pkg.CLTools_Executables"
+  else
+    chk_cmd="xcode-select -p"
+  fi
+  $chk_cmd 1>/dev/null 2>&1
+  if [ $? -ne 0 ]; then
+    # https://apple.stackexchange.com/questions/107307/how-can-i-install-the-command-line-tools-completely-from-the-command-line
+    touch /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress;
+    PROD=$(softwareupdate -l |
+    grep "\*.*Command Line.*$(sw_vers -productVersion|awk -F. '{print $1"."$2}')" |
+    head -n 1 | awk -F"*" '{print $2}' |
+    sed -e 's/^ *//' |
+    tr -d '\n')
+    softwareupdate -i "$PROD" --verbose
+    rm /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+  fi
 
   # install brew
   if [ ! -x "$(command -v brew)" ]; then
-    /usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+    URL_BREW='https://raw.githubusercontent.com/Homebrew/install/master/install'
+
+    echo -n '- Installing brew ... '
+    echo | /usr/bin/ruby -e "$(curl -fsSL $URL_BREW)" > /dev/null
+    if [ $? -eq 0 ]; then echo 'OK'; else echo 'NG'; fi
   fi
+fi
+
+if [ ! -f "$HOME/.profile" ] && [ ! -f "$HOME/.bash_profile" ] && [ ! -f "$HOME/.bash_login" ]; then
+  zetup link "$ZETUP_CUR_PKG/bash/profile.sh" "$HOME/.profile"
 fi
 
 brew_pkgs=(
   # your brew packages here
-  "wget"
+  "wget" # file downloader
+  "cask" # another package installer
+  "git" # source control (newer version)
+  "htop" # system monitor
+  "nmap" # security scanner
+  "links" # terminal web browser
+  "geoip" # geolocation of ip addresses
+  "irssi" # irc
+  "bash-completion"
+  "watch" # watch for changes in output of command
+  "tmux" # terminal multiplexer
 )
 brew_install ${brew_pkgs[@]}
+
+brew_cask_install_pkgs=(
+  "iterm2"
+)
+brew_cask_install ${brew_cask_install_pkgs}
 
 ssh-add $ZETUP_PRIVATE_KEY_FILE  2>/dev/null
 
@@ -97,18 +130,18 @@ source "$HOME/.bashrc"
 
 # uncomment to enable subpackages
 default_subpkgs_to_install=(
-  "chrome"
-  "docker"
-  "git"
-  "go"
-  #" keyboard-shortcuts" not ready
-  #"kubernetes"
-  "node"
-  #"ui"
-  #"video"
-  "vim"
-  "virtualbox"
-  "vscode"
+  # "chrome"
+  # "docker"
+  # "git"
+  # "go"
+  # #" keyboard-shortcuts" not ready
+  # #"kubernetes"
+  # "node"
+  # #"ui"
+  # #"video"
+  # "vim"
+  # "virtualbox"
+  # "vscode"
 )
 
 cache_subpkgs="$(zetup cache get subpkgs)"
